starting-kit/requirements.txt
matplotlib==3.5.2
scikit-image==0.19.3
pillow==9.2.0
torch==1.12.1
torchmetrics==0.10.0
pytorch_lightning==1.7.7
segmentation_models_pytorch==0.3.0
albumentations==1.2.1
rasterio==1.2.10
tensorboard==2.10.1
PyYAML==6.0
jupyterlab==3.5.0